requirements.txt
# 核心依赖
requests>=2.31.0
python-dotenv>=1.0.0
click>=8.1.0

# 开发和测试依赖
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# 可选依赖
rich>=13.0.0  # 美化命令行输出

# GUI依赖 (tkinter是Python内置的，无需额外安装)
# 如果在某些Linux发行版上tkinter不可用，可能需要安装:
# Ubuntu/Debian: sudo apt-get install python3-tk
# CentOS/RHEL: sudo yum install tkinter
# 或者: sudo dnf install python3-tkinter
